Fix to work with perls since 5.37.7
This fixes https://github.com/Perl/perl5/issues/20571

The commit blamed in that tickect caused POSIX::localeconv() to return
all the defined fields in the lconv structure.  Number::Format had been
relying on it to suppress empty fields.

The solution is simply for Number::Format to instead ignore empty fields
itself.  This is entirely backwards compatible, as previously those
fields were never delivered to Number::Format.

There is a slight complication in that the POSIX standard allows for
the monetary decimal point and the monetary thousands separator to both
be empty.  There is a check in Number::Format against both being empty
that must be relaxed to allow this legal combination.  The check
previously was effectively useless as Number::Format would never see
those fields if they were empty, as POSIX::localeconv() would suppress
them.